Add linter to travis
