+++ .golangci.yaml
@@ -1,0 +1,8 @@
+issues:
+  max-issues-per-linter: 0
+  max-same-issues: 0
+  exclude-rules:
+    - linters:
+      - errcheck
+      text: "Unsubscribe"
+      path: jsv2/jetstream/consumer.go
