+++ go_test.mod
@@ -7,6 +7,7 @@ require (
 	github.com/nats-io/nats-server/v2 v2.9.6
 	github.com/nats-io/nkeys v0.4.4
 	github.com/nats-io/nuid v1.0.1
+	golang.org/x/text v0.3.6
 	google.golang.org/protobuf v1.23.0
 )
