+++ jsv2/jetstream/stream_config.go
@@ -5,6 +5,9 @@ import (
 	"fmt"
 	"strings"
 	"time"
+
+	"golang.org/x/text/cases"
+	"golang.org/x/text/language"
 )
 
 type (
@@ -240,11 +243,12 @@ const (
 )
 
 func (st StorageType) String() string {
+	caser := cases.Title(language.AmericanEnglish)
 	switch st {
 	case MemoryStorage:
-		return strings.Title(memoryStorageString)
+		return caser.String(memoryStorageString)
 	case FileStorage:
-		return strings.Title(fileStorageString)
+		return caser.String(fileStorageString)
 	default:
 		return "Unknown Storage Type"
 	}
